refactor op_router to be split
